requirements.txt
pygame
screeninfo
pywin32
python-vlc